Improve keyword detection, remove old debug info
Config option added: ignoreSelf. If true, the bot will not trigger if
the bot's name contains the keyword and the message contains the bot's
name.

Setting the announce channel made a bit more robust, should now be
possible to start the bot and then join the announce channel (previously
this would cause the bot to crash after receiving a message). Untested.

The bot no longer triggers on anything said in the announce channel. If
this is something that might be useful in some cases this behavior could
be made optional.

Removed debug info when the bot received a message it didn't want to
respond to.